refactor: chat client
