+++ cortex-js/src/infrastructure/commanders/usecases/usecases.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { HttpModule } from '@nestjs/axios';
+import { FileManagerModule } from '@/infrastructure/services/file-manager/file-manager.module';
+import { PSCliUsecases } from './ps.cli.usecases';
+
+@Module({
+  imports: [HttpModule, FileManagerModule],
+  providers: [PSCliUsecases],
+  exports: [PSCliUsecases],
+})
+export class CliUsecasesModule {}
